+++ src/models/index.ts
@@ -2,3 +2,4 @@
  * API response models
  */
 export * from './responses';
+export * from './taxcloud';